chore: removed circular dep and css.escape
