+++ package.json
@@ -76,7 +76,6 @@
     "conventional-changelog-cli": "^2.0.25",
     "copy-dir": "^1.2.0",
     "cross-env": "^6.0.3",
-    "css.escape": "^1.5.1",
     "cssnano": "^4.1.10",
     "cypress": "^3.8.1",
     "cypress-image-snapshot": "^3.1.1",
