+++ packages/docsify-server-renderer/index.js
@@ -1,13 +1,13 @@
 import { readFileSync } from 'fs';
 import { resolve, basename } from 'path';
 import resolvePathname from 'resolve-pathname';
-import debug from 'debug';
-import fetch from 'node-fetch';
 import { AbstractHistory } from '../../src/core/router/history/abstract';
 import { Compiler } from '../../src/core/render/compiler';
 import { isAbsolutePath } from '../../src/core/router/util';
 import * as tpl from '../../src/core/render/tpl';
 import { prerenderEmbed } from '../../src/core/render/embed';
+import fetch from 'node-fetch';
+import debug from 'debug';
 
 function cwd(...args) {
   return resolve(process.cwd(), ...args);
