+++ src/core/global-api.js
@@ -1,10 +1,10 @@
-import marked from 'marked';
-import prism from 'prismjs';
 import * as util from './util';
 import * as dom from './util/dom';
 import { Compiler } from './render/compiler';
 import { slugify } from './render/slugify';
 import { get } from './fetch/ajax';
+import prism from 'prismjs';
+import marked from 'marked';
 
 export default function() {
   window.Docsify = {
