+++ src/core/render/index.js
@@ -1,5 +1,4 @@
 /* eslint-disable no-unused-vars */
-import tinydate from 'tinydate';
 import * as dom from '../util/dom';
 import cssVars from '../util/polyfill/css-vars';
 import { callHook } from '../init/lifecycle';
@@ -11,6 +10,7 @@ import { scrollActiveSidebar } from '../event/scroll';
 import { Compiler } from './compiler';
 import * as tpl from './tpl';
 import { prerenderEmbed } from './embed';
+import tinydate from 'tinydate';
 
 function executeScript() {
   const script = dom
