+++ src/core/render/utils.js
@@ -1,0 +1,20 @@
+export function getAndRemoveConfig(str = '') {
+  const config = {};
+
+  if (str) {
+    str = str
+      .replace(/^'/, '')
+      .replace(/'$/, '')
+      .replace(/(?:^|\s):([\w-]+:?)=?([\w-]+)?/g, (m, key, value) => {
+        if (key.indexOf(':') === -1) {
+          config[key] = (value && value.replace(/&quot;/g, '')) || true;
+          return '';
+        }
+
+        return m;
+      })
+      .trim();
+  }
+
+  return { str, config };
+}
