+++ test/unit/util.test.js
@@ -2,8 +2,8 @@
 require = require('esm')(
   module /* , options */
 ); /* eslint-disable-line no-global-assign */
-const { expect } = require('chai');
 const { resolvePath } = require('../../src/core/router/util');
+const { expect } = require('chai');
 
 describe('router/util', function() {
   it('resolvePath', async function() {
